Removed skip macro T.
T was essentially redundant. We can skip n bits using the read macro A
and just ignore the return value. Note that T(48); needed to be
changed into A(24);A(24); because the input buffer is only 32 bits
wide. There is still a significant reduction in character count as we
completely removed T's definition.